requirements.txt
langchain~=0.3.27
faiss-cpu~=1.11.0.post1
sentence-transformers~=5.0.0
torch~=2.7.1
transformers~=4.53.3
scikit-learn~=1.6.1